Minor fixes

+++ llmc/attention.cuh
@@ -227,7 +227,7 @@ void attention_forward(floatX* out, floatX* qkvr, floatX* att,
                                      B * NH, cublas_compute, CUBLAS_GEMM_DEFAULT));
 
     // multiply all elements of preatt elementwise by scale
-    float scale = use_mup ? 1.0f / HS : 1.f / sqrtf((float)HS);
+    float scale = use_mup ? 1.f / (float)HS : 1.f / sqrtf((float)HS);
     int grid_size = CEIL_DIV(B * NH * T * WARP_SIZE, block_size);
     softmax_forward_kernel5<<<grid_size, block_size, 0, stream>>>(att, scale, preatt, B * NH, T);
 
@@ -283,7 +283,7 @@ void attention_backward(floatX* dinp, floatX* dqkvr, floatX* datt, floatX* scrat
     cublasCheck(cublasGemmStridedBatchedEx(cublas_handle, CUBLAS_OP_N, CUBLAS_OP_T, HS, T, T, &alpha,
                                            scratch, CUBLAS_LOWP, HS, T * HS, att, CUBLAS_LOWP, T, T * T, &beta,
                                            dv, CUBLAS_LOWP, HS, T * HS, B * NH, cublas_compute, CUBLAS_GEMM_DEFAULT));
-    const float scale = use_mup ? 1.0f / HS : 1.f / sqrtf((float)HS);
+    const float scale = use_mup ? 1.f / (float)HS : 1.f / sqrtf((float)HS);
     // backward into preatt. this is an in-place operation; datt turns into dpreatt here
     softmax_autoregressive_backward_inplace_kernel<<<dim3(T / 4, B * NH), 256>>>(datt, att, B, T, C, scale);
     const floatX* dpreatt = datt;
